Adicionado erro ao deletar o acesso do criador da tarefa

+++ Services/Services/Implementation/AtividadeService.cs
@@ -81,6 +81,13 @@ namespace Services.Services.Implementation
                 throw new EntidadeNaoEncontrada("Atividade Funcionario");
             }
 
+
+            if(atividadeFuncionario.NivelAcesso == NivelAcesso.Todos)
+            {
+                throw new CriadorNaoPodeSerRemovido();
+            }
+
+
             _repository.AtividadeFuncionarioRepository.Delete(atividadeFuncionario);
             _repository.Save();
         }
